backend/requirements_upgraded.txt
fastapi==0.111.0
fastapi-cli==0.0.24
uvicorn==0.29.0
starlette==0.37.2
pydantic==2.7.1
python-dotenv==1.2.1
python-multipart==0.0.9
email-validator==2.3.0
httpx==0.27.0
requests==2.32.5
orjson==3.11.7
ujson==5.11.0
jinja2==3.1.6
numpy==2.4.2
scikit-learn==1.8.0
scipy==1.17.1
joblib==1.5.3
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
transformers==4.41.0
sentence-transformers==3.0.0
sentencepiece==0.2.0
tokenizers==0.19.1
safetensors==0.7.0
huggingface_hub==0.36.2
pdfplumber==0.11.0
pdfminer.six==20231228
pymupdf==1.24.3
pypdfium2==5.5.0
pillow==12.1.1
regex==2026.2.19
tqdm==4.67.3
pyyaml==6.0.3
rich==14.3.3
typer==0.24.1
python-pptx==0.6.23
python-docx==1.1.2
docx2txt==0.8
rouge-score==0.1.2
bert-score==0.3.13
PyMuPDF==1.24.3
faster-whisper==1.0.3
python-docx==1.1.2
reportlab==4.2.2
